Refactoring Code.
